Don't make it depend on dev-features
