+++ src/csb/import/mod.rs
@@ -1,4 +1,10 @@
 //! CSB import flow and related routes.
 mod pages;
 
+#[cfg(feature = "fixtures")]
+mod fixture;
+
+#[cfg(feature = "fixtures")]
+pub use fixture::import_csb_fixture;
+
 pub use pages::{CsbImportPath, router};
